+++ utils/handle_symbol.c
@@ -6,7 +6,7 @@
 /*   By: mgoudin <mgoudin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/03 12:05:30 by mgoudin           #+#    #+#             */
-/*   Updated: 2022/06/06 17:15:28 by mgoudin          ###   ########.fr       */
+/*   Updated: 2022/06/09 14:45:56 by mgoudin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,6 +20,20 @@ int	write_heredoc(char *heredoc, int fd)
     return (fd);
 }
 
+int is_space(char *str)
+{
+    int i;
+
+    i = 0;
+    while (str[i])
+    {
+        if (str[i] == ' ')
+            return (1);
+        i++;
+    }
+    return (0);
+}
+
 int	ft_heredoc(char *limiter)
 {
 	char	*line;
@@ -28,7 +42,10 @@ int	ft_heredoc(char *limiter)
 
     fd = open(".heredoc", O_CREAT | O_RDWR | O_TRUNC, 0644);
     if (fd < 0)
+    {
         print_error("Error\nFile Descriptor\n");
+        return (0);
+    }
     line = get_next_line(0);
 	if (ft_strncmp(line, limiter, ft_strlen(limiter)) == 0)
 		return (fd);
@@ -42,20 +59,84 @@ int	ft_heredoc(char *limiter)
     return (fd);
 }
 
+char    *replace_env_link(char *str)
+{
+    int i;
+    int j;
+    char *tmp;
+
+    i = 0;
+    j = 0;
+    while(str[i])
+    {
+        if (str[i] == '$')
+        {
+            j = 1;
+            while(str[i + j] && str[i + j] != ' ' && str[i + j] != '$')
+                j++;
+            tmp = getenv(ft_strn(&str[i + 1], j - 2));
+            if (tmp == 0)
+            {
+                str = replace_len(str, "", j);
+                i++;
+                continue;
+            }
+            if (is_space(tmp))
+            {
+                print_error("ambiguous redirect\n");
+                return (0);
+            }
+            str = replace_len(str, tmp, j);
+            i += ft_strlen(tmp) - 1;
+        }
+        i++;
+    }
+    return (str);
+}
+
 int    ft_redirect_out(t_list *lst)
 {
     int fd;
     char *arg;
+    char *tmp;
+    int type;
 
-    arg = (char *)((t_cmd *)lst->next->content)->content;
+    ((t_cmd *)lst->content)->type = delete;
+    lst = lst->next;
+    if (!lst)
+        return (0);
+    arg = replace_env_link((char *)((t_cmd *)lst->content)->content);
+    if (!arg)
+        return (0);
+    ((t_cmd *)lst->content)->type = delete;
+    lst = lst->next;
+    if (lst)
+        type = (int)((t_cmd *)lst->content)->type;
+    else
+        type = 0;
+    while(lst && (type == 8 || type == 9 || type == 10))
+    {
+        tmp = replace_env_link((char *)((t_cmd *)lst->content)->content);
+        if (!tmp)
+            return (0);
+        arg = ft_strjoin(arg, tmp);
+        ((t_cmd *)lst->content)->type = delete;
+        lst = lst->next;
+        if (lst)
+            type = (int)((t_cmd *)lst->content)->type;
+    }
     fd = open(arg, O_DIRECTORY | O_WRONLY, 0644);
     if (fd > 0)
+    {
         print_error(ft_strjoin(arg, " is a directory\n"));
+        return (0);
+    }
     fd = open(arg, O_WRONLY | O_CREAT, 0644);
     if (fd == -1)
+    {
         print_error(ft_strjoin(arg, " does not exist\n"));
-    ((t_cmd *)lst->content)->type = delete;
-    ((t_cmd *)lst->next->content)->type = delete;
+        return (0);
+    }
     return (fd);
 }
 
@@ -63,15 +144,41 @@ int    ft_redirect_in(t_list *lst)
 {
     int fd;
     char *arg;
+    char *tmp;
+    int type;
 
-    arg = (char *)((t_cmd *)lst->next->content)->content;
-    fd = open(arg, O_DIRECTORY | O_RDONLY | O_CREAT, 0644);
-    if (fd == -1)
-        fd = open(arg, O_RDONLY | O_CREAT, 0644);
-    if (fd == -1)    
-        print_error(ft_strjoin(arg, " does not exist\n"));
     ((t_cmd *)lst->content)->type = delete;
-    ((t_cmd *)lst->next->content)->type = delete;
+    lst = lst->next;
+    if (!lst)
+        return (0);
+    arg = replace_env_link((char *)((t_cmd *)lst->content)->content);
+    if (!arg)
+        return (0);
+    ((t_cmd *)lst->content)->type = delete;
+    lst = lst->next;
+    if (lst)
+        type = (int)((t_cmd *)lst->content)->type;
+    else
+        type = 0;
+    while(lst && (type == 8 || type == 9 || type == 10))
+    {
+        tmp = replace_env_link((char *)((t_cmd *)lst->content)->content);
+        if (!tmp)
+            return (0);
+        arg = ft_strjoin(arg, tmp);
+        ((t_cmd *)lst->content)->type = delete;
+        lst = lst->next;
+        if (lst)
+            type = (int)((t_cmd *)lst->content)->type;
+    }
+    fd = open(arg, O_DIRECTORY | O_RDONLY, 0644);
+    if (fd == -1)
+        fd = open(arg, O_RDONLY, 0644);
+    if (fd == -1)    
+    {
+        print_error(ft_strjoin(arg, " does not exist\n"));
+        return (0);
+    }
     return (fd);
 }
 
@@ -79,16 +186,45 @@ int    ft_doubleredirect_out(t_list *lst)
 {
     int fd;
     char *arg;
+    char *tmp;
+    int type;
 
-    arg = (char *)((t_cmd *)lst->next->content)->content;
+    ((t_cmd *)lst->content)->type = delete;
+    lst = lst->next;
+    if (!lst)
+        return (0);
+    arg = replace_env_link((char *)((t_cmd *)lst->content)->content);
+    if (!arg)
+        return (0);
+    ((t_cmd *)lst->content)->type = delete;
+    lst = lst->next;
+    if (lst)
+        type = (int)((t_cmd *)lst->content)->type;
+    else
+        type = 0;
+    while(lst && (type == 8 || type == 9 || type == 10))
+    {
+        tmp = replace_env_link((char *)((t_cmd *)lst->content)->content);
+        if (!tmp)
+            return (0);
+        arg = ft_strjoin(arg, tmp);
+        ((t_cmd *)lst->content)->type = delete;
+        lst = lst->next;
+        if (lst)
+            type = (int)((t_cmd *)lst->content)->type;
+    }
     fd = open(arg, O_DIRECTORY | O_WRONLY, 0644);
     if (fd > 0)
+    {
         print_error(ft_strjoin(arg, " is a directory\n"));
+        return (0);
+    }
     fd = open(arg, O_WRONLY | O_CREAT | O_APPEND, 0644);
     if (fd == -1)
+    {
         print_error(ft_strjoin(arg, " does not exist\n"));
-    ((t_cmd *)lst->content)->type = delete;
-    ((t_cmd *)lst->next->content)->type = delete;
+        return (0);
+    }
     return (fd);
 }
 
@@ -99,17 +235,22 @@ int    ft_doubleredirect_in(t_list *lst)
 
     arg = (char *)((t_cmd *)lst->next->content)->content;
     fd = ft_heredoc(arg);
+    if (!fd)
+        return (0);
     ((t_cmd *)lst->content)->type = delete;
     ((t_cmd *)lst->next->content)->type = delete;
     return (fd);
 }
 
-void ft_pipe(t_list *lst, t_redirect tab[], int j, int size)
+int ft_pipe(t_list *lst, t_redirect tab[], int j, int size)
 {
     int  pfd[2];
 
     if (pipe(pfd) == -1)
+    {
         print_error("pipe failed\n");
+        return (0);
+    }
     tab[j].out = pfd[1];
     tab[j].lst_pfd_in = pfd[1];
     tab[j].lst_pfd_out = pfd[0];
@@ -119,9 +260,10 @@ void ft_pipe(t_list *lst, t_redirect tab[], int j, int size)
         tab[j + 1].st_pfd_in = pfd[1];
         tab[j + 1].st_pfd_out = pfd[0];
     }
+    return (1);
 }
 
-t_redirect  *handle_symbol(t_list **head, t_list **head_symbol, int len)
+t_redirect  *handle_symbol(t_list **head, int len)
 {
     t_list *lst;
     int type;
@@ -148,19 +290,44 @@ t_redirect  *handle_symbol(t_list **head, t_list **head_symbol, int len)
         type = (int)((t_cmd *)lst->content)->type;
         if (type == pipe_)
         {
+            if (!(lst->next))
+            {
+                print_error("Nothing after last pipe\n");
+                return (NULL);
+            }
             if ((int)((t_cmd *)lst->next->content)->type == pipe_)
-                print_error("Two followed pipe");
-            ft_pipe(lst, tab, j, len);
+            {
+                print_error("Two followed pipe\n");
+                return (NULL);
+            }
+            if (!ft_pipe(lst, tab, j, len))
+                return (NULL);
             j++;
         }
         if (type == redirect_right)
+        {
             tab[j].out = ft_redirect_out(lst); 
+            if (tab[j].out == 0)
+                return (NULL);
+        }
         if (type == redirect_left)
+        {
             tab[j].in = ft_redirect_in(lst); 
+            if (tab[j].in == 0)
+                return (NULL);
+        }
         if (type == double_redirect_right)
+        {
             tab[j].out = ft_doubleredirect_out(lst); 
+            if (tab[j].out == 0)
+                return (NULL);
+        }
         if (type == double_redirect_left)
+        {
             tab[j].in = ft_doubleredirect_in(lst);
+            if (tab[j].in == 0)
+                return (NULL);
+        }
         lst = lst->next;
     }
     return (tab);
